% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmigen.R
\name{bmigen}
\alias{bmigen}
\title{An individual history simulator for BMI, GERD onset age, and death}
\usage{
bmigen(nsub = 100, sex = 1, race = 1, byrmin = 1800, byrmax = 2000,
  agemin = 12, agemax = 99, cyrmin = 1900, cyrmax = 2030)
}
\arguments{
\item{nsub}{Number of subjects.  Defaults to 100.}

\item{sex}{1 = males, 2 = females, defaults to 1}

\item{race}{1 = white, 2 = black, defaults to 1}

\item{byrmin}{Birth year range minimum defaults to 1800}

\item{byrmax}{Birth year range maximum defaults to 2000}

\item{agemin}{Age range minimum defaults to 12}

\item{agemax}{Age range maximum defaults to 99}

\item{cyrmin}{Calendar year range minimum defaults to 1900}

\item{cyrmax}{Calendar year range maximum defaults to 2030}
}
\description{
An individual history simulator for BMI, GERD onset age, and death
}
\examples{
bmigen(200)  #simulates 200 male Whites born any time from 1800-2000 including  GERD onset age, age at death, and annual BMI (from age 12 - 99). Writes to file: "BMIsimWM.csv"

bmigen(1000,1,2,1890,1900,,,,)  #simulates 1000 male Blacks born between 1890-1900 including  GERD onset age, age at death, and annual BMI (from age 12 - 99). Writes to file: "BMIsimBM.csv"

bmigen(1000,2,1,,,60,70,,)  #simulates 1000 female Whites  born between 1890-1900 including  GERD onset age, age at death, and annual BMI (from age 12 - 99). Writes to file: "BMIsimWF.csv"

bmigen(100,2,2,,,,,2000,2000)  #simulates 100 female Blacks with followup in year 2000 including  GERD onset age, age at death, and annual BMI (from age 12 - 99). Writes to file: "BMIsimBF.csv"


#File format:  "id", sex, race, birthyear, firstage, firstyear GERDonsetage deathage deathyear Number_BMIs BMIvalues

#Notes:  999 represents no GERD onset age or no age at death.  Annual simulated BMI values continue after simulated age at death. 

}
\keyword{BMI}
\keyword{Death}
\keyword{GERD}
\keyword{NHANES}

